Refactor: repository 인터페이스, 구현부 분리
