+++ app/src/main/java/com/sopt/now/compose/data/repository/FriendRepository.kt
@@ -1,13 +1,7 @@
 package com.sopt.now.compose.data.repository
 
-import com.sopt.now.compose.data.service.FriendService
 import com.sopt.now.compose.presentation.friend.Friend
 
-class FriendRepository(private val friendService: FriendService) {
-    suspend fun getFriendList(page: Int): Result<List<Friend>> {
-        return runCatching {
-            val response = friendService.getFriendList(page)
-            response.data
-        }
-    }
+interface FriendRepository {
+    suspend fun getFriendList(page: Int): Result<List<Friend>>
 }
